ai/requirements.txt
# ================================================================================
# 핵심 딥러닝 프레임워크
# ================================================================================

# PyTorch (CUDA 12.1)
--extra-index-url https://download.pytorch.org/whl/cu121
torch>=2.5.1
torchvision>=0.20.1
torchaudio>=2.5.1

# Transformers & HuggingFace
huggingface-hub>=0.35.3
transformers>=4.57.1
accelerate>=1.10.1
bitsandbytes>=0.48.1
sentencepiece>=0.2.1
tokenizers>=0.22.1
safetensors>=0.6.2

# FastAPI
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pydantic>=2.0.0

# 유튜브 도구
yt-dlp>=2023.11.0
youtube-transcript-api>=0.6.1

# 영상 도구
opencv-python>=4.8.0
scenedetect[opencv]>=0.6.2
ffmpeg-python>=0.2.0

# Whisper (로컬 음성 인식)
openai-whisper>=20231117

# Llama 3.2 Vision 모델용 추가 의존성
protobuf

# 유틸리티
python-dotenv>=1.0.0
pillow>=12.0.0
requests>=2.32.5
numpy>=2.3.4
tqdm>=4.67.1
psutil>=7.1.0
PyYAML>=6.0.3

# 의존성 패키지
Jinja2>=3.1.6
MarkupSafe>=3.0.3
certifi>=2025.10.5
charset-normalizer>=3.4.4
idna>=3.11
urllib3>=2.5.0
filelock>=3.20.0
fsspec>=2025.9.0
packaging>=25.0
regex>=2025.9.18
typing_extensions>=4.15.0
colorama>=0.4.6
networkx>=3.5
mpmath>=1.3.0
sympy>=1.13.1
setuptools>=80.9.0
